+++ test/unit/events/guild-member-add/index.test.ts
@@ -1,5 +1,5 @@
 import * as GuildMemberAddModule from '../../../../src/events/guild-member-add';
-import { SendWelcomeMessageHandler } from '../../../../src/events/guild-member-add/handlers';
+import { SendWelcomeMessageHandler } from '../../../../src/events/guild-member-add/handlers/send-welcome-message';
 
 
 describe('GuildMemberAddsModule', () => {
