+++ test/unit/events/message-reaction-remove/handlers/revoke-role.test.ts
@@ -1,8 +1,8 @@
 import sinon, { StubbedInstance, stubInterface, stubObject } from 'ts-sinon';
-import { should } from 'chai';
 import { Message, MessageReaction, User, GuildMember, Role } from 'discord.js';
+import { should } from 'chai';
 
-import { RevokeRoleHandler } from '../../../../../src/events/message-reaction-remove';
+import { RevokeRoleHandler } from '../../../../../src/events/message-reaction-remove/handlers/revoke-role';
 import { Logger } from '../../../../../src/lib/logger';
 import { ExtractedMessageProps } from '../../../../../src/types';
